Fix URL value column
